modernise ethers
