+++ apps/laboratory/src/components/Ethers/EthersTransactionTest.tsx
@@ -1,0 +1,68 @@
+import { Button, useToast, Stack, Link, Text, Spacer } from '@chakra-ui/react'
+import { useWeb3ModalAccount, useWeb3ModalProvider } from '@web3modal/ethers/react'
+import { BrowserProvider, JsonRpcSigner, ethers } from 'ethers'
+import { vitalikEthAddress } from '../../utils/DataUtil'
+import { sepolia } from '../../utils/ChainsUtil'
+import { useState } from 'react'
+
+export function EthersTransactionTest() {
+  const toast = useToast()
+  const { address, chainId } = useWeb3ModalAccount()
+  const { walletProvider } = useWeb3ModalProvider()
+  const [loading, setLoading] = useState(false)
+
+  async function onSendTransaction() {
+    try {
+      setLoading(true)
+      if (!walletProvider || !address) {
+        throw Error('user is disconnected')
+      }
+      const provider = new BrowserProvider(walletProvider, chainId)
+      const signer = new JsonRpcSigner(provider, address)
+      const tx = await signer.sendTransaction({
+        to: vitalikEthAddress,
+        value: ethers.parseUnits('0.001', 'gwei')
+      })
+      toast({ title: 'Succcess', description: tx.blockHash, status: 'success', isClosable: true })
+    } catch {
+      toast({
+        title: 'Error',
+        description: 'Failed to sign transaction',
+        status: 'error',
+        isClosable: true
+      })
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  return chainId === sepolia.chainId && address ? (
+    <Stack direction={['column', 'column', 'row']}>
+      <Button
+        data-test-id="sign-transaction-button"
+        onClick={onSendTransaction}
+        isDisabled={loading}
+      >
+        Send Transaction to Vitalik
+      </Button>
+
+      <Spacer />
+
+      <Link isExternal href="https://sepoliafaucet.com">
+        <Button variant="outline" colorScheme="blue" isDisabled={loading}>
+          Sepolia Faucet 1
+        </Button>
+      </Link>
+
+      <Link isExternal href="https://www.infura.io/faucet/sepolia">
+        <Button variant="outline" colorScheme="orange" isDisabled={loading}>
+          Sepolia Faucet 2
+        </Button>
+      </Link>
+    </Stack>
+  ) : (
+    <Text fontSize="md" color="yellow">
+      Switch to Sepolia Ethereum Testnet to test this feature
+    </Text>
+  )
+}
